Move `from_sync` and `from_async` implementations to single file
